Build pwd output using repo name and subfolder
If working directory was, say, /var/www/html and it was a git repo,
the path would not be replaced by the repo name, as it expected the
folder name to be in the path (~/dev/repos/basherk for basherk repo).

After refactoring, we add the subfolder within the repo (show-prefix)
directly to the repo name, avoiding string replace functions.

We also handle some edge cases, and manually build subfolder name for
the .git folder, as show-prefix doesn't think it's inside the repo.