Remove inoperable args parsing from f
When searching file contents with f, we detect which of rg/ag/grep to
use, in that order.
Originally all other arguments passed to f would be passed through to
the relevant tool. `f in test -i` would map to `grep "test" -i ...`

Due to incorrect quoting, this functionality regressed in 0b2347e.

Furthermore, since the fallback to grep is meant for compatibility on
barebones servers without rg/ag installed, assuming perl is installed
(for the args regex) is a mistake, and it regularly isn't.