Handle symlinks better in tab function
If you cd into a symlinked folder, and cd ../foo, bash knows you want
to go back to the previous folder, and then traverse to foo, but when
inside the symlink, [[ -d ../foo ]] will usually fail as it looks for
foo inside physical parent.
To handle this better in tab, we cd into the folder in a subshell, and
return the path if successful, allowing cd to traverse the symlink tree
if necessary.