Reorganise aliases & hoisted functions in header
Reorganised the top portion of basherk with headings that are described
in README.md > Contributing.

Removed some old aliases that haven't been useful in a long time, but
most code was moved/re-ordered rather than changed.

Migrated osver and suho to functions to more simplify OS-specific alias
logic in the header.